web/src/utils/Validator.ts
import isEmpty from 'lodash/isEmpty';

type Value = any;

const pattern = {
  // http://emailregex.com/
  url: /^(?!mailto:)(?:(?:http|https|ftp):\/\/|\/\/)(?:\S+(?::\S*)?@)?(?:(?:(?:[1-9]\d?|1\d\d|2[01]\d|22[0-3])(?:\.(?:1?\d{1,2}|2[0-4]\d|25[0-5])){2}(?:\.(?:[0-9]\d?|1\d\d|2[0-4]\d|25[0-4]))|(?:(?:[a-z\u00a1-\uffff0-9]+-*)*[a-z\u00a1-\uffff0-9]+)(?:\.(?:[a-z\u00a1-\uffff0-9]+-*)*[a-z\u00a1-\uffff0-9]+)*(?:\.(?:[a-z\u00a1-\uffff]{2,})))|localhost)(?::\d{2,5})?(?:(\/|\?|#)[^\s]*)?$/i,
};

const Validator = {
  required(value: Value) {
    return !isEmpty(value);
  },
  url(value: Value) {
    return typeof value === 'string' && value.length <= 2048 && Boolean(value.match(pattern.url));
  },
};

export default Validator;
